Refed textures

+++ src/texture.js
@@ -17,7 +17,14 @@ class Texture {
     } = settings;
     if (!output) throw new Error('settings property "output" required.');
     if (!context) throw new Error('settings property "context" required.');
+    if (!texture) throw new Error('settings property "texture" required.');
+    if (!kernel) throw new Error('settings property "kernel" required.');
     this.texture = texture;
+    if (texture.refs) {
+      texture.refs++;
+    } else {
+      texture.refs = 1;
+    }
     this.size = size;
     this.dimensions = dimensions;
     this.output = output;
@@ -52,7 +59,12 @@ class Texture {
    * @desc Deletes the Texture
    */
   delete() {
+    if (this._deleted) return;
     this._deleted = true;
+    if (this.texture.refs) {
+      this.texture.refs--;
+      if (this.texture.refs) return;
+    }
     return this.context.deleteTexture(this.texture);
   }
 }
